Add "GAIA-X" context(s)
Ass described in issue 200, the latest VCs out of the wizard can not be
added to the catalog, since the used context is unknown and will not be
loaded for security reasons. Additional contexts can be configured now
and will be added to the VPs during verification.

Issue: https://gitlab.eclipse.org/eclipse/xfsc/cat/fc-service/-/issues/200