Update on: "fix: undo change in neo4j Java versions"
The gds version 2.5.2 did not work proper in an embedded neo4j used by
unit tests. Version 2.4.6 seems to do the work. All JUnit tests pass
now.
In neo4j version 5.7 an error message within an exception was changed
(message about a timeout). This was updated in the corresponding class.
Hope there are not other breaking changes not caught by a unit test :-(